Feat: 스파서비스 리뷰 상세 조회 기능 및 예약 엔티티(+테이블)에도 price 컬럼 추가 fix : 취소된 예약도 재예약되도록, price 타입 수정(int -> Long)

+++ src/main/java/com/example/oyl/dto/ReservationSummaryDTO.java
@@ -23,12 +23,12 @@ public class ReservationSummaryDTO {
     private LocalTime reservationTime;    // 예약 시간
     private ReservationStatus reservationStatus;        // "RESERVED" or "CANCELED"
     private RefundStatus refundStatus;                  // "NONE", "WAITING", "COMPLETED", "REJECTED"
-    private int price;
+    private Long price;
 
     // 생성자 명시적으로 추가해주기
     public ReservationSummaryDTO(String reservationId, String userNickname, String dogName, String dogId,
                                  String serviceName, String serviceId, LocalDate reservationDate, LocalTime reservationTime,
-                                 ReservationStatus reservationStatus, RefundStatus refundStatus, int price) {
+                                 ReservationStatus reservationStatus, RefundStatus refundStatus, Long price) {
         this.reservationId = reservationId;
         this.userNickname = userNickname;
         this.dogName = dogName;
